Extract ActionView::TemplatePath
