+++ actionview/lib/action_view/testing/resolvers.rb
@@ -36,7 +36,7 @@ module ActionView #:nodoc:
 
   class NullResolver < Resolver
     def find_templates(name, prefix, partial, details, locals = [])
-      path = Path.build(name, prefix, partial)
+      path = TemplatePath.build(name, prefix, partial)
       handler = ActionView::Template::Handlers::Raw
       [ActionView::Template.new("Template generated by Null Resolver", path.virtual, handler, virtual_path: path.virtual, format: nil, variant: nil, locals: locals)]
     end
